task/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
PyJWT==2.3.0
simplejson==3.17.6
